protocol/x/lending/keeper/keeper.go
package keeper

import (
	"fmt"

	cosmoslog "cosmossdk.io/log"
	storetypes "cosmossdk.io/store/types"
	"github.com/StreamFinance-Protocol/stream-chain/protocol/indexer/indexer_manager"
	"github.com/StreamFinance-Protocol/stream-chain/protocol/x/lending/types"
	"github.com/cosmos/cosmos-sdk/codec"
	sdk "github.com/cosmos/cosmos-sdk/types"
)

type (
	Keeper struct {
		cdc                 codec.BinaryCodec
		storeKey            storetypes.StoreKey
		indexerEventManager indexer_manager.IndexerEventManager
	}
)

func NewKeeper(
	cdc codec.BinaryCodec,
	storeKey storetypes.StoreKey,
	indexerEventManager indexer_manager.IndexerEventManager,
) *Keeper {
	return &Keeper{
		cdc:                 cdc,
		storeKey:            storeKey,
		indexerEventManager: indexerEventManager,
	}
}

func (k Keeper) Logger(ctx sdk.Context) cosmoslog.Logger {
	return ctx.Logger().With(cosmoslog.ModuleKey, fmt.Sprintf("x/%s", types.ModuleName))
}

func (k Keeper) InitializeForGenesis(ctx sdk.Context) {}

/////////// Lending Accounts //////////

// marshals and stores the lending account in kv-store
func (k Keeper) SetLendingAccount(ctx sdk.Context, account types.LendingAccount) {
	store := ctx.KVStore(k.storeKey)
	accountKey := types.GetLendingAccountStoreKey(account.Address)

	bz := k.cdc.MustMarshal(&account)
	store.Set(accountKey, bz)
}

// Retrieves the lending account from the KVStore.
func (k Keeper) GetLendingAccount(ctx sdk.Context, bech32AccAddr string) (types.LendingAccount, bool) {
	store := ctx.KVStore(k.storeKey)
	accountKey := types.GetLendingAccountStoreKey(bech32AccAddr)
	if !store.Has(accountKey) {
		return types.LendingAccount{}, false
	}
	bz := store.Get(accountKey)
	var account types.LendingAccount
	k.cdc.MustUnmarshal(bz, &account)

	// Ensure accountPositions field is initialized to an empty slice if it's nil
	if account.AccountPositions == nil {
		account.AccountPositions = []*types.AccountPosition{}
	}

	return account, true
}

// ///////// Lending Pools //////////
func (k Keeper) SetPool(ctx sdk.Context, pool types.Pool) {
	store := ctx.KVStore(k.storeKey)
	bz := k.cdc.MustMarshal(&pool)
	store.Set([]byte(pool.TotalDeposits.Denom), bz)
}

func (k Keeper) GetPool(ctx sdk.Context, assetDenom string) (types.Pool, bool) {
	store := ctx.KVStore(k.storeKey)
	bz := store.Get([]byte(assetDenom))
	if bz == nil {
		return types.Pool{}, false
	}
	var pool types.Pool
	k.cdc.MustUnmarshal(bz, &pool)
	return pool, true
}

// Core functionalities

// Borrow
func (k Keeper) Borrow(ctx sdk.Context, borrower sdk.AccAddress, amount sdk.Coin) error {
	// Implementation to borrow assets
	return nil
}

func (k Keeper) Repay(ctx sdk.Context, borrower sdk.AccAddress, amount sdk.Coin) error {
	// Implementation to repay borrowed assets
	return nil
}

func (k Keeper) Withdraw(ctx sdk.Context, depositor sdk.AccAddress, amount sdk.Coin) error {
	// Implementation to withdraw assets
	return nil
}

// Interest accrual and collateral management
func (k Keeper) AccrueInterest(ctx sdk.Context) {
	// Implementation to accrue interest on deposits and loans
}

func (k Keeper) DepositCollateral(ctx sdk.Context, depositor sdk.AccAddress, collateral sdk.Coin) error {
	// Implementation to deposit collateral
	return nil
}

func (k Keeper) WithdrawCollateral(ctx sdk.Context, depositor sdk.AccAddress, collateral sdk.Coin) error {
	// Implementation to withdraw collateral
	return nil
}

// Liquidation
func (k Keeper) Liquidate(ctx sdk.Context, borrower sdk.AccAddress, liquidator sdk.AccAddress, amount sdk.Coin) error {
	// Implementation to liquidate undercollateralized loans
	return nil
}

// Governance and parameters
// func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
// 	// Implementation to set parameters
// }

// func (k Keeper) GetParams(ctx sdk.Context) types.Params {
// 	// Implementation to get parameters
// 	//return types.Params{}
// }

// Event emission
func (k Keeper) EmitDepositEvent(ctx sdk.Context, depositor sdk.AccAddress, amount sdk.Coin) {
	// Implementation to emit deposit event
}

func (k Keeper) EmitBorrowEvent(ctx sdk.Context, borrower sdk.AccAddress, amount sdk.Coin) {
	// Implementation to emit borrow event
}

// Helper functions
func (k Keeper) getAccount(ctx sdk.Context, address sdk.AccAddress) types.LendingAccount {
	// Implementation to get account details
	return types.LendingAccount{}
}

func (k Keeper) setAccount(ctx sdk.Context, account types.LendingAccount) {
	// Implementation to set account details
}
